backend/requirements.txt
Django>=4.2.7,<5.0
djangorestframework>=3.14.0,<4.0
djangorestframework-simplejwt>=5.3.0,<6.0
django-cors-headers>=4.3.1,<5.0
gunicorn>=21.0.0
whitenoise>=6.5.0
psycopg2-binary>=2.9.0
dj-database-url>=2.0.0
openai>=1.99.0
python-dotenv>=1.0.0
requests>=2.31.0
